+++ classes/scripts/runner.php
@@ -651,13 +651,11 @@ class runner extends atoum\script\configurable
 
                             exit(3);
                         }
-                    }
-                        );
+                    });
 
                     try {
                         $score = $autorunner->run()->getRunner()->getScore();
-
-                        $isSuccess = $score->getFailNumber() <= 0 && $score->getErrorNumber() <= 0 && $score->getExceptionNumber() <= 0;
+                        $isSuccess = $score->getFailNumber() <= 0 && $score->getErrorNumber() <= 0 && $score->getExceptionNumber() <= 0 && $score->getUncompletedMethodNumber() <= 0;
 
                         if ($autorunner->shouldFailIfVoidMethods() && $score->getVoidMethodNumber() > 0) {
                             $isSuccess = false;